requirements.txt
aiohttp==3.9.3
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.6.0
anyio==3.7.1
APScheduler==3.10.4
attrs==21.4.0
bcrypt==4.1.2
beautifulsoup4==4.12.2
bs4==0.0.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==42.0.5
dnspython==2.5.0
ecdsa==0.18.0
email-validator==2.1.0.post1
fastapi==0.109.2
frozenlist==1.4.1
greenlet==3.0.3
grobid==0.2.0
grobid-client==0.8.5
grobid-client-python==0.0.8
grobid2json==0.0.1
h11==0.12.0
httpcore==0.15.0
httptools==0.6.1
httpx==0.23.0
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
lxml==4.9.4
MarkupSafe==2.1.5
multidict==6.0.5
orjson==3.9.14
passlib==1.7.4
pillow==10.3.0
pyasn1==0.5.1
pycparser==2.21
pydantic==2.6.1
pydantic-extra-types==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.16.2
PyJWT==2.8.0
PyMuPDF==1.24.2
PyMuPDFb==1.24.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-jose==3.3.0
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
requests==2.31.0
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
SQLAlchemy==2.0.29
starlette==0.36.3
twilio==8.13.0
typing_extensions==4.9.0
tzdata==2024.1
tzlocal==5.2
ujson==5.9.0
urllib3==2.2.1
uvicorn==0.27.1
watchfiles==0.21.0
websockets==12.0
yarl==1.9.4
